Improve ComponentSemantics API and eliminate allocation overhead
- Add Deref/DerefMut on ComponentScoping → ComponentSemantics, removing
  ~25 pure forwarding methods
- Replace 10 FxHashSet<SymbolId> boolean classifications with per-symbol
  state bits in Vec<u32> (hash lookup → array index + bitmask)
- Add name_index to SymbolTable for O(1) find_symbol_by_name (was O(n))
- Move has_write_reference_other_than to ComponentSemantics
- Add generic state bit API (has_symbol_state/set_symbol_state/symbols_with_state)
- Widen symbol state from u8 to u32 (bits 0-7 core, 8+ for consumers)
- Export sym_state module for core state bit constants

Net: -79 lines, 5000 fewer heap allocations per large project build,
better cache locality for symbol classification queries in codegen.

+++ crates/svelte_analyze/src/validate/runes.rs
@@ -366,7 +366,7 @@ struct StateRefLocallyValidator<'a, 'b> {
     /// Incremented when entering `$derived`/`$derived.by` call arguments.
     /// Mirrors the reference compiler's `function_depth += 1` for `$derived` calls:
     /// references inside `$derived(...)` are semantically deeper and should not warn.
-    derived_call_depth: usize,
+    derived_call_depth: u32,
     _phantom: std::marker::PhantomData<&'a ()>,
 }
